pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wealth-sim-germany"
version = "0.1.0"
description = "Monte Carlo simulation of Germany's wealth distribution and public finance."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "UNLICENSED" }
authors = [{ name = "Economy Simulator Contributors" }]
keywords = ["simulation", "economics", "germany", "wealth"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.6.0",
  "mypy>=1.10",
  "types-PyYAML>=6.0.12",
]

[tool.setuptools.packages.find]
include = ["wealth_sim_germany"]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["B", "C4", "E", "F", "I", "RUF", "SIM", "UP"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
ignore_missing_imports = true
